Убрал лишнии стуктуры данных
